Gate all the optimizations under preprocessor options

+++ command/test.sh
@@ -18,6 +18,24 @@ if [[ "$OSTYPE" == "darwin"* ]]; then
     options="$options $macos_suppress_options"
 fi
 
-$CC tests.c optiscope.c -o tests $options
-./tests
-rm tests
+# Run all the combinations of optimizations, exhaustively.
+first=1
+for a in "" "-DOPTISCOPE_DISABLE_DELIMITER_COMPRESSION"; do
+for b in "" "-DOPTISCOPE_DISABLE_DELIMITER_SCHEDULING"; do
+for c in "" "-DOPTISCOPE_DISABLE_DELIMITER_EXTRUSION"; do
+for d in "" "-DOPTISCOPE_DISABLE_CLOSEDNESS_ANNOTATIONS"; do
+for e in "" "-DOPTISCOPE_DISABLE_SEGMENTATION"; do
+    if [ -n "$d" ] && [ -z "$e" ]; then continue; fi
+    if [ -n "$a" ] && [ -z "$b" ]; then continue; fi
+    if [ "$first" -eq 0 ]; then echo ""; fi
+    first=0
+    extra=$(echo $a $b $c $d $e)
+    if [ -z "$extra" ]; then echo "<empty>"; else echo "$extra"; fi
+    $CC tests.c optiscope.c -o tests $options $extra
+    ./tests
+    rm tests
+done
+done
+done
+done
+done
